requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
openai>=1.30.0
